==> 01_convert_wav.ipynb <==
import os

from utils.audio_control import extract_audio

video_list = os.listdir('video')

for video_name in video_list:
    video_path = os.path.join('video', video_name)
    extract_audio(video_path, 'audio')
