==> 05_split_csv.ipynb <==
import os
import pandas as pd

class FileName:
    TRAJECTORIES_NAME = "trajectories.csv"

TRAJECTORIES_DIR = "trajectories"
os.makedirs(TRAJECTORIES_DIR, exist_ok=True)

csv_dir = "csv"
csv_list = [os.path.join(csv_dir, f) for f in os.listdir(csv_dir) if f.endswith(".csv")]
csv_list = sorted(csv_list)

def main(csv_path):
    with open(csv_path, 'r') as f:
        file_name = os.path.basename(csv_path).split(".")[0]
        TRAJECTORIES_PATH = f"{TRAJECTORIES_DIR}/{file_name}_{FileName.TRAJECTORIES_NAME}"
        if os.path.exists(TRAJECTORIES_PATH):
            return None

        lines = f.readlines()
        lines_iter = iter(lines)
        while True:
            try:
                line = next(lines_iter).strip().split(',')
                if line[0] == 'TRAJECTORIES':
                    # Hz
                    line = next(lines_iter).strip().split(',')
                    TRAJECTORIES_Hz = float(line[0])
                    # columns 1
                    line = next(lines_iter).strip().split(',')
                    TRAJECTORIES_COLUMNS_1 = line
                    # columns 2
                    line = next(lines_iter).strip().split(',')
                    TRAJECTORIES_COLUMNS_2 = line
                    # data
                    TRAJECTORIES_DATA = pd.DataFrame(columns=TRAJECTORIES_COLUMNS_2)
                    count_index = 0
                    while True:
                        try:
                            count_index += 1
                            line = next(lines_iter).strip().split(',')
                            index = int(line[0]) if line[0].isdigit() else line[0]
                            if index != count_index:
                                # organize data
                                TRAJECTORIES_COLUMNS_1 = [x for x in TRAJECTORIES_COLUMNS_1 if x != '']
                                TRAJECTORIES_DATA.set_index(TRAJECTORIES_COLUMNS_2[0], inplace=True)
                                NEW_COLUMNS = [f"{col1}_{axis}" for col1 in TRAJECTORIES_COLUMNS_1 for axis in ['x', 'y', 'z']]
                                TRAJECTORIES_DATA.columns = NEW_COLUMNS
                                TRAJECTORIES_DATA.to_csv(TRAJECTORIES_PATH)
                                break
                            TRAJECTORIES_DATA.loc[index] = line
                        except StopIteration:
                            break
                    pass
            except StopIteration:
                break

for i, csv_path in enumerate(csv_list):
    try:
        print(f"{i+1}/{len(csv_list)}: {csv_path}")
        main(csv_path)
    except Exception as e:
        print(f"Error: {csv_path}")
        continue
