==> 07_fix_trajectory.ipynb <==
import os
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

from IPython.display import display

pd.set_option('display.max_columns', 500)

fps = 250
thr = 2

output_dir = 'fixed_trajectories'

for num in range(10):
    trajectory_path = f"trajectories/hirasaki_{num}_trajectories.csv"
    
    df = pd.read_csv(trajectory_path, index_col=0)
    keypoints_columns = [col.replace("_x", "") for col in df.columns if col.endswith("_x") and "*" not in col]
    drop_columns = [col for col in df.columns if "*" in col]
    
    null_rows = df[[col + "_x" for col in keypoints_columns]].isnull().sum(axis=1).values
    max_index = np.argmax(null_rows > thr)
    min_index = np.argmax(null_rows == 0)
    
    df = df.iloc[min_index:max_index]
    df = df.drop(columns=drop_columns)
    
    df = df.interpolate(method="spline", order=3, limit_direction="both")
    df.to_csv(os.path.join(output_dir, trajectory_path.split("/")[-1]))
