==> 09_check_toe_off.ipynb <==
import os

import cv2
from PIL import Image

import numpy as np
import pandas as pd

import matplotlib.pyplot as plt

df = pd.read_csv("camera_sync.csv")
df.head()

for data_num in range(10):
    toe_off_0 = df["toe_off_0"].iloc[data_num]
    toe_off_1 = df["toe_off_1"].iloc[data_num]
    toe_off_2 = df["toe_off_2"].iloc[data_num]
    toe_off_3 = df["toe_off_3"].iloc[data_num]
    toe_off_4 = df["toe_off_4"].iloc[data_num]

    min_frame = min(toe_off_0, toe_off_1, toe_off_2, toe_off_3, toe_off_4)

    frame0 = toe_off_0 - min_frame
    frame1 = toe_off_1 - min_frame
    frame2 = toe_off_2 - min_frame
    frame3 = toe_off_3 - min_frame
    frame4 = toe_off_4 - min_frame

    fps = 20
    output_path = f"output/hirasaki_walk_sync_{data_num}.mp4"
    fourcc = cv2.VideoWriter_fourcc(*"mp4v")
    video = None

    while True:
        frame0 += 1
        frame1 += 1
        frame2 += 1
        frame3 += 1
        frame4 += 1
            
        image_path_1 = f"img/hirasaki_0_{data_num}/hirasaki_0_{data_num}_{frame0}.jpg"
        image_path_2 = f"img/hirasaki_1_{data_num}/hirasaki_1_{data_num}_{frame1}.jpg"
        image_path_3 = f"img/hirasaki_2_{data_num}/hirasaki_2_{data_num}_{frame2}.jpg"
        image_path_4 = f"img/hirasaki_3_{data_num}/hirasaki_3_{data_num}_{frame3}.jpg"
        image_path_5 = f"img/hirasaki_4_{data_num}/hirasaki_4_{data_num}_{frame4}.jpg"

        image1 = cv2.imread(image_path_1)
        image2 = cv2.imread(image_path_2)
        image3 = cv2.imread(image_path_3)
        image4 = cv2.imread(image_path_4)
        image5 = cv2.imread(image_path_5)
        
        if image1 is None or image2 is None or image3 is None or image4 is None or image5 is None:
            break
        
        image31 = cv2.hconcat([image1, image2])
        image32 = cv2.hconcat([image3, image4])
        image33 = cv2.hconcat([image5, np.zeros_like(image5)])
        image = cv2.vconcat([image31, image32, image33])
        
        if video is None:
            height, width, _ = image.shape
            video = cv2.VideoWriter(output_path, fourcc, fps, (width, height))
        
        video.write(image)

    video.release()
    cv2.destroyAllWindows()
