==> 12_calculate_center_trajectory.ipynb <==
import os

import pandas as pd

from utils.skeleton_keypoints import *

input_dir = "fixed_trajectories"
output_dir = "fixed_trajectories_2"

os.makedirs(output_dir, exist_ok=True)

input_files = [os.path.join(input_dir, f) for f in os.listdir(input_dir) if f.endswith(".csv")]
input_files.sort()

for input_file in input_files:
    df = pd.read_csv(input_file, index_col=0)
    
    df2 = pd.DataFrame()
    for key, value in keypoint_pairs_for_calc.items():
        col1, col2 = keypoint_pairs_for_calc[key]
        
        df2[f"{key}_x"] = (df[f"{col1}_x"] + df[f"{col2}_x"]) / 2
        df2[f"{key}_y"] = (df[f"{col1}_y"] + df[f"{col2}_y"]) / 2
        df2[f"{key}_z"] = (df[f"{col1}_z"] + df[f"{col2}_z"]) / 2

    df2.index = df.index
    df2.to_csv(os.path.join(output_dir, os.path.basename(input_file)))
